third_party/mahcinarium/gdb: mm gdb support
There are many situations in odyssey that must be debugged.
But now it is really hard to do it because of machinarium coroutines engine.

This patch series will add gdb support for machinarium's
coroutines in golang's goroutines style.

For the first step command 'info mmcoros' was added, and now it
can be used to display the list of running coroutines for each thread (mm_machine).

See gdb/README.md for more.

Signed-off-by: Roman Khapov <r.khapov@ya.ru>