Remove client notify io (#698)
* sources/client: remove notify_io

The only purpose of it is to send kill command
to the clients. But using of eventfd for that is overkill and creates
an 2x usage of fd for odyssey.

This patch replaces the eventfd to atomic var.

Signed-off-by: rkhapov <r.khapov@ya.ru>

* client.h: remove useless signal

Signed-off-by: rkhapov <r.khapov@ya.ru>

---------

Signed-off-by: rkhapov <r.khapov@ya.ru>
Co-authored-by: rkhapov <r.khapov@ya.ru>
(cherry picked from commit 4fde8dbad4ca7b30e2f50ac0592b511e99e4e8a4)
Signed-off-by: rkhapov <r.khapov@ya.ru>