Msg leaks somewhere (#742)
* machinarium/channels: add get_size funcs

Can be used in some debuggin workarounds

Signed-off-by: rkhapov <r.khapov@ya.ru>

* machinarium/wait_list*

This patch adds mm_wait_list_t type
which will be used as condition variable everywhere it needed.

Ex.: wait_bus in router creates it's own wait_list, but with channels
and therefore it has memory leaks because of inconsistency of it usages.

Signed-off-by: rkhapov <r.khapov@ya.ru>

* sources/route.h: fix wait_bus inconsistency

route->wait_bus was implemented with channels.
This leads to memory leaks because of inconsystency of queued
clients notifications.

This patch replaces channel with wait_list, which will not lead
to memory leaks.

Signed-off-by: rkhapov <r.khapov@ya.ru>

* sources/config_reader.c: fix small memleak

just to silence asan

Signed-off-by: rkhapov <r.khapov@ya.ru>

---------

Signed-off-by: rkhapov <r.khapov@ya.ru>
Co-authored-by: rkhapov <r.khapov@ya.ru>
(cherry picked from commit 806856e27069d3667090286928cd275d43f06610)
Signed-off-by: rkhapov <r.khapov@ya.ru>