sources/machine.c: up stack size by three times
There is a problem when odyssey is compiled with openssl >= 3.0.0

The problem is SIGSEGV inside openssl functions because of
stackoverflow. This happens because odyssey allocates thread with
minimum possible stack size, which works with openssl 1.1.1, but leads
to SIGSEGV on openssl >= 3

This patch increases machine's stack size by three times to solve the
issue.

Signed-off-by: rkhapov <r.khapov@ya.ru>