Build with prometheus-client-c library (#944)
* docker/prometheus-legacy: added build with prometheus-client-c in docker

The original library was not working. This commit make a patch for it and fix some issues with the cron.c file when working with this library.

* CMakeLists.txt: Changed work with the prometheus-client-c library.

Changed the start format for the prom lib in the makefile and Dockerfile.

* docker/prometheus-legacy: add build with prometheus

Added test for prometheus metrics collection, fixed bug with odyssey's daemonization

* Fix format

* Added prometheus build to workflows